Projet_API-Youssef-Medkouri/pyproject.toml
[tool.poetry]
name = "restaurant"
version = "0.1.0"
description = ""
authors = ["Adam Hoffstein <adam@C0377-Macbook-Pro.local>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
django-rest-registration = "^0.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
